Turn off autoupdate and more bug-fixing
I am using AutoHotkey_2.0-a100-52515e2. When I did the latest update, lot of the script stopped working because of the considerable changes in array/object syntax.

I dont want to change these till the new syntax is somewhat finalized. So AHK should not be updated anymore for the time being.